+++ includes/minishell.h
@@ -6,7 +6,7 @@
 /*   By: emadriga <emadriga@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/07 00:43:55 by jvacaris          #+#    #+#             */
-/*   Updated: 2022/01/15 17:52:27 by emadriga         ###   ########.fr       */
+/*   Updated: 2022/01/16 13:43:46 by emadriga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -102,7 +102,8 @@ void	ft_unset(t_str **env_list, char **argv);
 char	*close_quotes_pipedfork(char *str_got_old);
 
 //*		forks / heredoc_handler.c
-t_str	*get_heredoc_list(char **input);
+void	get_heredoc_list(char **input, t_str **heredoc_list);
+char	*get_heredoc_pipedfork(const char *key);
 
 //		old / exec_command.c
 //// void	exec_command(char **list, char **envp);
@@ -219,5 +220,7 @@ char	*new_getpath(char *raw_cmd, t_str **env_list);
 
 //*		var_expansor.c
 char	*recursive_expand(char *malloc_str, t_str **env_list);
+
+//*		piped_processes.c
 void	get_piped_processes(char **tokens, t_pp **processes);
 #endif